Rename `generation.py` to `text_generation.py`
