+++ tests/integration/test_pipe_llms.py
@@ -20,7 +20,7 @@ from distilabel.llm.openai import OpenAILLM
 from distilabel.pipeline.local import Pipeline
 from distilabel.pipeline.step.base import Step
 from distilabel.pipeline.step.generators.huggingface import LoadHubDataset
-from distilabel.pipeline.step.task.generation import TextGeneration
+from distilabel.pipeline.step.task.text_generation import TextGeneration
 from distilabel.pipeline.step.typing import RuntimeParameter, StepInput
